% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{ansett}
\alias{ansett}
\title{Passenger numbers on Ansett airline flights}
\format{
Time series of class \code{tsibble}
}
\source{
Ansett Airlines (which no longer exists).
}
\description{
The data features a major pilots' industrial dispute which results in some
weeks having zero passengers. There were also at least two changes in the
definitions of passenger classes.
}
\details{
\code{ansett} is a weekly \code{tsibble} with one value:
\tabular{ll}{
Passengers:  \tab Total air passengers travelling with Ansett\cr
}

Each series is uniquely identified using two keys:
\tabular{ll}{
Airports:  \tab The airports that passengers are travelling between (both directions)\cr
Class:     \tab The class of the ticket.\cr
}
}
\examples{
library(tsibble)
ansett

}
\keyword{datasets}
