% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aus_production}
\alias{aus_production}
\title{Quarterly production of selected commodities in Australia.}
\format{
Time series of class \code{tsibble}.
}
\source{
Australian Bureau of Statistics, catalogue number 8301.0.55.001 table 1.
}
\description{
Quarterly estimates of selected indicators of manufacturing production in Australia.
}
\details{
\code{aus_production} is a half-hourly \code{tsibble} with six values:
\tabular{ll}{
Beer:       \tab Beer production in megalitres.\cr
Tobacco:    \tab Tobacco and cigarette production in tonnes.\cr
Bricks:     \tab Clay brick production in millions of bricks.\cr
Cement:     \tab Portland cement production in thousands of tonnes.\cr
Electricity:\tab Electricity production in gigawatt hours.\cr
Gas:        \tab Gas production in petajoules.\cr
}
}
\examples{
library(tsibble)
aus_production

}
\keyword{datasets}
