% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{aus_retail}
\alias{aus_retail}
\title{Australian retail trade turnover}
\format{Time series of class \code{tsibble}}
\source{
Australian Bureau of Statistics, catalogue number 8501.0, table 11.
}
\description{
\code{aus_retail} is a monthly \code{tsibble} with one value:
\tabular{ll}{
Turnover:      \tab Retail turnover in $Million AUD\cr
}
}
\details{
Each series is uniquely identified using two keys:
\tabular{ll}{
State:     \tab The Australian state (or territory)\cr
Industry:     \tab The industry of retail trade\cr
}
}
\keyword{datasets}
