% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecastingdata_org.R
\name{convert_tsf_to_tsibble}
\alias{convert_tsf_to_tsibble}
\title{Convert .tsf data into tsibble format}
\usage{
convert_tsf_to_tsibble(
  file,
  value_column_name = "series_value",
  key = NULL,
  index = NULL
)
}
\arguments{
\item{file}{.tsf file path}

\item{value_column_name}{Any name that is preferred to have as the name of the column containing series values in the returning tsibble.}

\item{key}{The name of the attribute that should be used as the key when creating the tsibble. If doesn't provide, a data frame will be returned instead of a tsibble.}

\item{index}{The name of the time attribute that should be used as the index when creating the tsibble. If doesn't provide, it will search for a valid index. When no valid index found, a data frame will be returned instead of a tsibble.}
}
\value{
This function returns a tsibble or a data.frame based on the validity of given key and index values.
}
\description{
This function converts the contents in a .tsf file into a tsibble or a data.frame and returns it along with other meta-data of the dataset: frequency, horizon, whether the dataset contains missing values and whether the series have equal lengths.
For more details on the tsf file format, please refer to: Godahewa, R., Bergmeir, C., Webb, G. I., Hyndman, R. J. & Montero-Manso, P. (2021), Monash Time Series Forecasting Archive.
}
\references{
Godahewa, R., Bergmeir, C., Webb, G. I., Hyndman, R. J. & Montero-Manso, P. (2021), Monash Time Series Forecasting Repository. \url{https://forecastingdata.org/}

Godahewa, R., Bergmeir, C., Webb, G. I., Hyndman, R. J. & Montero-Manso, P. (2021), Monash Time Series Forecasting Archive.
}
