% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecastingdata_org.R
\name{download_data_zenodo}
\alias{download_data_zenodo}
\title{Download data from the Monash Forecasting Repository}
\usage{
download_data_zenodo(
  record_id,
  file_name,
  index = NULL,
  key = "series_name",
  value_column_name = "series_value"
)
}
\arguments{
\item{record_id}{Record identifier of the file that needs to be downloaded from the Monash Time Series Forecasting Repository}

\item{file_name}{Name of the file (without extensions) that needs to be downloaded from the Monash Time Series Forecasting Repository}

\item{index}{The name of the time attribute that should be used as the index when creating the tsibble. If doesn't provide, it tries to find a valid index within the data. If there is no valid index, then a data frame will be returned instead of a tsibble.}

\item{key}{The name of the attribute that should be used as the key when creating the tsibble. The default value is "series_name" which is the key of the datasets in our repository. If doesn't provide, a data frame will be returned instead of a tsibble.}

\item{value_column_name}{Any name that is preferred to have as the name of the column containing series values in the returning tsibble.}
}
\value{
This function returns a tsibble or a data.frame based on the validity of given key and index values.
}
\description{
This function downloads a zip file from the Monash Time Series Forecasting Repository given the record identifier and file name, unzips the contents and converts the unzipped .tsf file into a tsibble or a data.frame given the key and index.
For more details on the repository and the tsf file format, please refer to: Godahewa, R., Bergmeir, C., Webb, G. I., Hyndman, R. J. & Montero-Manso, P. (2021), Monash Time Series Forecasting Archive.
}
\references{
Godahewa, R., Bergmeir, C., Webb, G. I., Hyndman, R. J. & Montero-Manso, P. (2021), Monash Time Series Forecasting Repository. \url{https://forecastingdata.org/}

Godahewa, R., Bergmeir, C., Webb, G. I., Hyndman, R. J. & Montero-Manso, P. (2021), Monash Time Series Forecasting Archive.
}
