% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{global_economy}
\alias{global_economy}
\title{Global economic indicators}
\format{
Time series of class \code{tsibble}
}
\source{
The World Bank, \url{http://datatopics.worldbank.org/world-development-indicators/}
}
\description{
Economic indicators featured by the World Bank from 1960 to 2017.
}
\details{
\code{global_economy} is an annual \code{tsibble} with six values:
\tabular{ll}{
GDP:       \tab Gross domestic product (in $USD February 2019).\cr
Growth:    \tab Annual percentage growth in GDP.\cr
CPI:       \tab Consumer price index (base year 2010).\cr
Imports:   \tab Imports of goods and services (\% of GDP).\cr
Exports:   \tab Exports of goods and services (\% of GDP).\cr
Population:\tab Total population.
}

Each series is uniquely identified by one key:
\tabular{ll}{
Country:   \tab The country or region of the series.\cr
}
}
\examples{
library(tsibble)
global_economy
}
\keyword{datasets}
