% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{PBS}
\alias{PBS}
\title{Monthly Medicare Australia prescription data}
\format{Time series of class \code{tsibble}}
\source{
Medicare Australia
}
\description{
\code{PBS} is a monthly \code{tsibble} with two values:
\tabular{ll}{
Scripts:      \tab Total number of scripts\cr
Cost:     \tab Cost of the scripts in $AUD\cr
}
}
\details{
The data is disaggregated using four keys:
\tabular{ll}{
Concession:      \tab Concessional scripts are given to pensioners, unemployed, dependents, and other card holders \cr
Type:      \tab Co-payments are made until an individual's script expenditure hits a threshold ($290.00 for concession, $1141.80 otherwise). Safety net subsidies are provided to individuals exceeding this amount. \cr
ATC1:     \tab Anatomical Therapeutic Chemical index (level 1)\cr
ATC2:     \tab Anatomical Therapeutic Chemical index (level 2)\cr
}
}
\examples{

PBS

}
\keyword{datasets}
