% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{pelt}
\alias{pelt}
\title{Pelt trading records}
\format{Time series of class \code{tsibble}}
\source{
Hudson Bay Company
}
\description{
Hudson Bay Company trading records for Snowshoe Hare and Canadian Lynx furs
from 1845 to 1935. This data contains trade records for all areas of the company.
}
\details{
\code{pelt} is an annual \code{tsibble} with two values:
\tabular{ll}{
Hare:      \tab The number of Snowshoe Hare pelts traded.\cr
Lynx:      \tab The number of Canadian Lynx pelts traded.\cr
}
}
\keyword{datasets}
