% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsibbledata.R
\docType{package}
\name{tsibbledata-package}
\alias{tsibbledata}
\alias{tsibbledata-package}
\title{tsibbledata: Diverse Datasets for tsibble}
\description{
This data package provides many datasets in the tsibble format which are suitable for learning and testing the tsibble and its extension packages.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tidyverts/tsibbledata}
  \item Report bugs at \url{https://github.com/tidyverts/tsibbledata/issues}
}

}
\author{
\strong{Maintainer}: Mitchell O'Hara-Wild \email{mail@mitchelloharawild.com}

Authors:
\itemize{
  \item Rob Hyndman
  \item Earo Wang
}

}
\keyword{internal}
