% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{aus_livestock}
\alias{aus_livestock}
\title{Australian livestock slaughter}
\format{Time series of class \code{tsibble}}
\source{
Australian Bureau of Statistics, catalogue number 7218.0.55.001 tables 1 to 6.
}
\description{
Meat production in Australia for human consumption from Q3 1965 to Q4 2018.
}
\details{
\code{aus_livestock} is an quarterly \code{tsibble} with one value:
\tabular{ll}{
Count:    \tab Number of animals slaughtered.\cr
}

Each series is uniquely identified by one key:
\tabular{ll}{
Animal:   \tab The animal slaughtered.\cr
}
}
\keyword{datasets}
