% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsibbledata.R
\docType{package}
\name{tsibbledata-package}
\alias{tsibbledata}
\alias{tsibbledata-package}
\title{tsibbledata: Example Datasets for tsibble}
\description{
This data package provides datasets in the correct structure for learning and testing the tidyverts family of packages.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tidyverts/tsibbledata}
  \item Report bugs at \url{https://github.com/tidyverts/tsibbledata/issues}
}

}
\author{
\strong{Maintainer}: Rob Hyndman \email{Rob.Hyndman@monash.edu} [copyright holder]

Authors:
\itemize{
  \item Mitchell O'Hara-Wild
  \item Earo Wang
}

}
\keyword{internal}
