% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsibbledata.R
\docType{package}
\name{tsibbledata-package}
\alias{tsibbledata}
\alias{tsibbledata-package}
\title{tsibbledata: Diverse Datasets for tsibble}
\description{
\if{html}{\figure{logo.png}{options: align='right'}}

This data package provides diverse datasets in the tsibble data structure. These datasets are useful for learning how tsibble can be applied on a variety of time series tasks.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tidyverts/tsibbledata}
  \item Report bugs at \url{https://github.com/tidyverts/tsibbledata/issues}
}

}
\author{
\strong{Maintainer}: Mitchell O'Hara-Wild \email{mail@mitchelloharawild.com}

Authors:
\itemize{
  \item Rob Hyndman
  \item Earo Wang
}

}
\keyword{internal}
