% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsibbledata.R
\docType{package}
\name{tsibbledata-package}
\alias{tsibbledata}
\alias{tsibbledata-package}
\title{tsibbledata: Diverse Datasets for 'tsibble'}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Provides diverse datasets in the 'tsibble' data structure. These datasets are useful for learning and demonstrating how tidy temporal data can tidied, visualised, and forecasted.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://tsibbledata.tidyverts.org/}
  \item \url{https://github.com/tidyverts/tsibbledata/}
  \item Report bugs at \url{https://github.com/tidyverts/tsibbledata/issues}
}

}
\author{
\strong{Maintainer}: Mitchell O'Hara-Wild \email{mail@mitchelloharawild.com}

Authors:
\itemize{
  \item Rob Hyndman
  \item Earo Wang
}

}
\keyword{internal}
